+++ web/modules/custom/giv_din_stemme/src/Controller/GivDinStemmeController.php
@@ -284,9 +284,8 @@ class GivDinStemmeController extends ControllerBase {
     // Load GivDinStemme.
     $givDinStemme = $this->helper->getGivDinStemmeByCollectionIdAndDelta($collection_id, $delta);
 
-    $metadata = json_decode($givDinStemme->get('metadata')->getValue()[0]['value'], TRUE);
-    $metadata['duration'] = $request->request->get('duration');
-    $givDinStemme->set('metadata', json_encode($metadata));
+    $this->helper->updateTotalDonationDuration((int) $request->request->get('duration'));
+    $this->helper->updateTotalNumberOfDonations();
 
     $files = $request->files->all();