+++ web/modules/custom/giv_din_stemme/src/Helper/Helper.php
@@ -4,6 +4,8 @@ namespace Drupal\giv_din_stemme\Helper;
 
 use Drupal\Component\Uuid\Php;
 use Drupal\Core\Entity\EntityTypeManagerInterface;
+use Drupal\Core\Lock\LockBackendInterface;
+use Drupal\Core\State\State;
 use Drupal\Core\StringTranslation\StringTranslationTrait;
 use Drupal\giv_din_stemme\Entity\GivDinStemme;
 use Drupal\giv_din_stemme\Exception\NoTextFoundException;
@@ -17,16 +19,37 @@ class Helper {
   use StringTranslationTrait;
 
   /**
+   * Donation count state key.
+   */
+  const GIV_DIN_STEMME_DONATION_COUNT_STATE_KEY = 'giv_din_stemme_donation_count';
+
+  /**
+   * Total donation duration (in seconds) state key.
+   */
+  const GIV_DIN_STEMME_TOTAL_DONATION_DURATION_STATE_KEY = 'giv_din_stemme_total_donation_duration';
+
+  /**
+   * Lock name for updating states.
+   */
+  const GIV_DIN_STEMME_UPDATE_STATE_LOCK = 'giv_din_stemme_update_state_lock';
+
+  /**
    * Constructor.
    *
    * @param \Drupal\Core\Entity\EntityTypeManagerInterface $entityTypeManager
    *   Entity type manager.
    * @param \Drupal\Component\Uuid\Php $uuid
    *   Uuid generator.
+   * @param \Drupal\Core\State\State $state
+   *   State system.
+   * @param \Drupal\Core\Lock\LockBackendInterface $lock
+   *   Lock manager.
    */
   public function __construct(
     protected EntityTypeManagerInterface $entityTypeManager,
     protected Php $uuid,
+    protected State $state,
+    protected LockBackendInterface $lock,
   ) {
   }
 
@@ -37,11 +60,65 @@ class Helper {
    *   Array of frontpage values
    */
   public function getFrontpageValues(): array {
-    // @todo when the entity is settled.
     return [
-      'donations' => 15,
-      'minutes' => 43,
+      'donations' => $this->getTotalNumberOfDonations() ?? 0,
+      'minutes' => $this->getTotalDonationDuration() ? ceil($this->getTotalDonationDuration() / 60) : 0,
     ];
+  }
+
+  /**
+   * Gets total donation duration.
+   */
+  private function getTotalDonationDuration(): ?int {
+    $value = $this->state->get(self::GIV_DIN_STEMME_TOTAL_DONATION_DURATION_STATE_KEY);
+
+    return $value ? (int) $value : NULL;
+  }
+
+  /**
+   * Update total donation duration state.
+   */
+  public function updateTotalDonationDuration(int $duration): void {
+
+    $this->getLock();
+
+    $totalDonationDuration = $this->getTotalDonationDuration() ?? 0;
+    $this->state->set(self::GIV_DIN_STEMME_TOTAL_DONATION_DURATION_STATE_KEY, $totalDonationDuration + $duration);
+
+    $this->lock->release(self::GIV_DIN_STEMME_UPDATE_STATE_LOCK);
+  }
+
+  /**
+   * Gets total number of donations.
+   */
+  private function getTotalNumberOfDonations(): ?int {
+    $value = $this->state->get(self::GIV_DIN_STEMME_DONATION_COUNT_STATE_KEY);
+
+    return $value ? (int) $value : NULL;
+  }
+
+  /**
+   * Adds one to total number of donations.
+   */
+  public function updateTotalNumberOfDonations(): void {
+
+    $this->getLock();
+
+    $currentTotal = $this->getTotalNumberOfDonations() ?? 0;
+    $this->state->set(self::GIV_DIN_STEMME_DONATION_COUNT_STATE_KEY, $currentTotal + 1);
+
+    $this->lock->release(self::GIV_DIN_STEMME_UPDATE_STATE_LOCK);
+  }
+
+  /**
+   * Gets the GIV_DIN_STEMME_UPDATE_STATE_LOCK lock.
+   */
+  private function getLock(): void {
+    // Attempt acquiring lock, wait if failed.
+    // @see https://api.drupal.org/api/drupal/core%21lib%21Drupal%21Core%21Lock%21LockBackendInterface.php/group/lock/11.x
+    while (!$this->lock->acquire(self::GIV_DIN_STEMME_UPDATE_STATE_LOCK)) {
+      $this->lock->wait(self::GIV_DIN_STEMME_UPDATE_STATE_LOCK);
+    }
   }
 
   /**