+++ CHANGELOG.md
@@ -8,6 +8,8 @@ See [keep a changelog] for information about writing changes to this log.
 
 ## [Unreleased]
 
+* [PR-24](https://github.com/itk-dev/giv-din-stemme/pull/24)
+  Keep total donation count and duration in state.
 * [PR-28](https://github.com/itk-dev/giv-din-stemme/pull/28)
   Ensure file is passed to server.
 * [PR-25](https://github.com/itk-dev/giv-din-stemme/pull/25)